Fix ZipSlip bug found by LGTM.com
The unsanitized path of a zip archive entry, which may
contain '..', was used directly to resolve the destination
path for the files being unzipped.

Although the prefix of the path was checked against `sourceFolder`, there could
be ".." path segments after that.

Extracting files from a malicious archive without validating
that the destination file path is within the destination
directory can cause files outside the destination directory
to be overwritten.